refactor: remove duplicate workplace home surface
